Fix of import of reminders with a date, on non-US systems.
Apple Script expects the date to follow the format set in the regional 
settings. 
Without this fix the app was crashing as described in issue #3
Change the code to use localizedStringFromDate fix the issue on my system
(I have locale set to a custom format).

I guess it will work for other locales too, but I haven't test it.